server, test: Refactor NodeCluster and ServerCluster (#14512)
ref tikv/tikv#14401

Parameterize NodeCluster and ServerCluster
