+++ components/test_raftstore-v2/src/server.rs
@@ -12,7 +12,7 @@ use causal_ts::CausalTsProviderImpl;
 use collections::{HashMap, HashSet};
 use concurrency_manager::ConcurrencyManager;
 use encryption_export::DataKeyManager;
-use engine_rocks::{RocksEngine, RocksSnapshot};
+use engine_rocks::RocksEngine;
 use engine_test::raft::RaftTestEngine;
 use engine_traits::{KvEngine, RaftEngine, TabletRegistry};
 use futures::{executor::block_on, Future};
@@ -86,23 +86,26 @@ impl FlowStatsReporter for DummyReporter {
     fn report_write_stats(&self, _write_stats: WriteStats) {}
 }
 
-type SimulateRaftExtension = <TestRaftKv2 as Engine>::RaftExtension;
-type SimulateStoreTransport = SimulateTransport<RaftRouter<RocksEngine, RaftTestEngine>>;
-type SimulateServerTransport =
-    SimulateTransport<ServerTransport<SimulateRaftExtension, PdStoreAddrResolver>>;
+type SimulateRaftExtension<EK> = <TestRaftKv2<EK> as Engine>::RaftExtension;
+type SimulateStoreTransport<EK> = SimulateTransport<RaftRouter<EK, RaftTestEngine>>;
+type SimulateServerTransport<EK> =
+    SimulateTransport<ServerTransport<SimulateRaftExtension<EK>, PdStoreAddrResolver>>;
 
-pub type SimulateEngine = RaftKv2<RocksEngine, RaftTestEngine>;
+pub type SimulateEngine<EK> = RaftKv2<EK, RaftTestEngine>;
 
 // TestRaftKvv2 behaves the same way with RaftKv2, except that it has filters
 // that can mock various network conditions.
 #[derive(Clone)]
-pub struct TestRaftKv2 {
-    raftkv: SimulateEngine,
+pub struct TestRaftKv2<EK: KvEngine> {
+    raftkv: SimulateEngine<EK>,
     filters: Arc<RwLock<Vec<Box<dyn Filter>>>>,
 }
 
-impl TestRaftKv2 {
-    pub fn new(raftkv: SimulateEngine, filters: Arc<RwLock<Vec<Box<dyn Filter>>>>) -> TestRaftKv2 {
+impl<EK: KvEngine> TestRaftKv2<EK> {
+    pub fn new(
+        raftkv: SimulateEngine<EK>,
+        filters: Arc<RwLock<Vec<Box<dyn Filter>>>>,
+    ) -> TestRaftKv2<EK> {
         TestRaftKv2 { raftkv, filters }
     }
 
@@ -111,15 +114,15 @@ impl TestRaftKv2 {
     }
 }
 
-impl Engine for TestRaftKv2 {
-    type Snap = RegionSnapshot<<RocksEngine as KvEngine>::Snapshot>;
-    type Local = RocksEngine;
+impl<EK: KvEngine> Engine for TestRaftKv2<EK> {
+    type Snap = RegionSnapshot<EK::Snapshot>;
+    type Local = EK;
 
     fn kv_engine(&self) -> Option<Self::Local> {
         self.raftkv.kv_engine()
     }
 
-    type RaftExtension = TestExtension;
+    type RaftExtension = TestExtension<EK>;
     fn raft_extension(&self) -> Self::RaftExtension {
         TestExtension::new(self.raftkv.raft_extension(), self.filters.clone())
     }
@@ -131,12 +134,12 @@ impl Engine for TestRaftKv2 {
         self.raftkv.modify_on_kv_engine(region_modifies)
     }
 
-    type SnapshotRes = <SimulateEngine as Engine>::SnapshotRes;
+    type SnapshotRes = <SimulateEngine<EK> as Engine>::SnapshotRes;
     fn async_snapshot(&mut self, ctx: SnapContext<'_>) -> Self::SnapshotRes {
         self.raftkv.async_snapshot(ctx)
     }
 
-    type WriteRes = <SimulateEngine as Engine>::WriteRes;
+    type WriteRes = <SimulateEngine<EK> as Engine>::WriteRes;
     fn async_write(
         &self,
         ctx: &Context,
@@ -159,21 +162,21 @@ impl Engine for TestRaftKv2 {
 }
 
 #[derive(Clone)]
-pub struct TestExtension {
-    extension: Extension<RocksEngine, RaftTestEngine>,
+pub struct TestExtension<EK: KvEngine> {
+    extension: Extension<EK, RaftTestEngine>,
     filters: Arc<RwLock<Vec<Box<dyn Filter>>>>,
 }
 
-impl TestExtension {
+impl<EK: KvEngine> TestExtension<EK> {
     pub fn new(
-        extension: Extension<RocksEngine, RaftTestEngine>,
+        extension: Extension<EK, RaftTestEngine>,
         filters: Arc<RwLock<Vec<Box<dyn Filter>>>>,
     ) -> Self {
         TestExtension { extension, filters }
     }
 }
 
-impl RaftExtension for TestExtension {
+impl<EK: KvEngine> RaftExtension for TestExtension<EK> {
     fn feed(&self, msg: RaftMessage, key_message: bool) {
         let send = |msg| -> raftstore::Result<()> {
             self.extension.feed(msg, key_message);
@@ -236,22 +239,22 @@ impl RaftExtension for TestExtension {
     }
 }
 
-pub struct ServerMeta {
-    node: NodeV2<TestPdClient, RocksEngine, RaftTestEngine>,
-    server: Server<PdStoreAddrResolver, TestRaftKv2>,
-    sim_router: SimulateStoreTransport,
-    sim_trans: SimulateServerTransport,
-    raw_router: StoreRouter<RocksEngine, RaftTestEngine>,
-    gc_worker: GcWorker<TestRaftKv2>,
+pub struct ServerMeta<EK: KvEngine> {
+    node: NodeV2<TestPdClient, EK, RaftTestEngine>,
+    server: Server<PdStoreAddrResolver, TestRaftKv2<EK>>,
+    sim_router: SimulateStoreTransport<EK>,
+    sim_trans: SimulateServerTransport<EK>,
+    raw_router: StoreRouter<EK, RaftTestEngine>,
+    gc_worker: GcWorker<TestRaftKv2<EK>>,
     rsmeter_cleanup: Box<dyn FnOnce()>,
 }
 
 type PendingServices = Vec<Box<dyn Fn() -> Service>>;
 
-pub struct ServerCluster {
-    metas: HashMap<u64, ServerMeta>,
+pub struct ServerCluster<EK: KvEngine> {
+    metas: HashMap<u64, ServerMeta<EK>>,
     addrs: AddressMap,
-    pub storages: HashMap<u64, TestRaftKv2>,
+    pub storages: HashMap<u64, TestRaftKv2<EK>>,
     pub region_info_accessors: HashMap<u64, RegionInfoAccessor>,
     snap_paths: HashMap<u64, TempDir>,
     snap_mgrs: HashMap<u64, TabletSnapManager>,
@@ -266,8 +269,8 @@ pub struct ServerCluster {
     pub causal_ts_providers: HashMap<u64, Arc<CausalTsProviderImpl>>,
 }
 
-impl ServerCluster {
-    pub fn new(pd_client: Arc<TestPdClient>) -> ServerCluster {
+impl<EK: KvEngine> ServerCluster<EK> {
+    pub fn new(pd_client: Arc<TestPdClient>) -> Self {
         let env = Arc::new(
             EnvBuilder::new()
                 .cq_count(2)
@@ -316,10 +319,10 @@ impl ServerCluster {
         &mut self,
         node_id: u64,
         mut cfg: Config,
-        store_meta: Arc<Mutex<StoreMeta<RocksEngine>>>,
+        store_meta: Arc<Mutex<StoreMeta<EK>>>,
         key_manager: Option<Arc<DataKeyManager>>,
         raft_engine: RaftTestEngine,
-        tablet_registry: TabletRegistry<RocksEngine>,
+        tablet_registry: TabletRegistry<EK>,
         resource_manager: &Option<Arc<ResourceGroupManager>>,
     ) -> ServerResult<u64> {
         let (snap_mgr, snap_mgs_path) = if !self.snap_mgrs.contains_key(&node_id) {
@@ -639,7 +642,7 @@ impl ServerCluster {
         Ok(node_id)
     }
 
-    pub fn get_gc_worker(&self, node_id: u64) -> &GcWorker<TestRaftKv2> {
+    pub fn get_gc_worker(&self, node_id: u64) -> &GcWorker<TestRaftKv2<EK>> {
         &self.metas.get(&node_id).unwrap().gc_worker
     }
 
@@ -677,7 +680,7 @@ impl ServerCluster {
     }
 }
 
-impl Simulator<RocksEngine> for ServerCluster {
+impl<EK: KvEngine> Simulator<EK> for ServerCluster<EK> {
     fn get_node_ids(&self) -> HashSet<u64> {
         self.metas.keys().cloned().collect()
     }
@@ -718,10 +721,10 @@ impl Simulator<RocksEngine> for ServerCluster {
         &mut self,
         node_id: u64,
         cfg: Config,
-        store_meta: Arc<Mutex<StoreMeta<RocksEngine>>>,
+        store_meta: Arc<Mutex<StoreMeta<EK>>>,
         key_manager: Option<Arc<DataKeyManager>>,
         raft_engine: RaftTestEngine,
-        tablet_registry: TabletRegistry<RocksEngine>,
+        tablet_registry: TabletRegistry<EK>,
         resource_manager: &Option<Arc<ResourceGroupManager>>,
     ) -> ServerResult<u64> {
         dispatch_api_version!(
@@ -754,9 +757,8 @@ impl Simulator<RocksEngine> for ServerCluster {
     fn async_snapshot(
         &mut self,
         request: kvproto::raft_cmdpb::RaftCmdRequest,
-    ) -> impl Future<
-        Output = std::result::Result<RegionSnapshot<engine_rocks::RocksSnapshot>, RaftCmdResponse>,
-    > + Send {
+    ) -> impl Future<Output = std::result::Result<RegionSnapshot<EK::Snapshot>, RaftCmdResponse>> + Send
+    {
         let node_id = request.get_header().get_peer().get_store_id();
         let mut router = match self.metas.get(&node_id) {
             None => {
@@ -792,7 +794,7 @@ impl Simulator<RocksEngine> for ServerCluster {
         Ok(())
     }
 
-    fn get_router(&self, node_id: u64) -> Option<StoreRouter<RocksEngine, RaftTestEngine>> {
+    fn get_router(&self, node_id: u64) -> Option<StoreRouter<EK, RaftTestEngine>> {
         self.metas.get(&node_id).map(|m| m.raw_router.clone())
     }
 
@@ -805,9 +807,9 @@ impl Simulator<RocksEngine> for ServerCluster {
     }
 }
 
-impl Cluster<ServerCluster, RocksEngine> {
-    pub fn must_get_snapshot_of_region(&mut self, region_id: u64) -> RegionSnapshot<RocksSnapshot> {
-        let mut try_snapshot = || -> Option<RegionSnapshot<RocksSnapshot>> {
+impl<EK: KvEngine> Cluster<ServerCluster<EK>, EK> {
+    pub fn must_get_snapshot_of_region(&mut self, region_id: u64) -> RegionSnapshot<EK::Snapshot> {
+        let mut try_snapshot = || -> Option<RegionSnapshot<EK::Snapshot>> {
             let leader = self.leader_of_region(region_id)?;
             let store_id = leader.store_id;
             let epoch = self.get_region_epoch(region_id);
@@ -833,7 +835,10 @@ impl Cluster<ServerCluster, RocksEngine> {
     }
 }
 
-pub fn new_server_cluster(id: u64, count: usize) -> Cluster<ServerCluster, RocksEngine> {
+pub fn new_server_cluster(
+    id: u64,
+    count: usize,
+) -> Cluster<ServerCluster<RocksEngine>, RocksEngine> {
     let pd_client = Arc::new(TestPdClient::new(id, false));
     let sim = Arc::new(RwLock::new(ServerCluster::new(Arc::clone(&pd_client))));
     Cluster::new(
@@ -849,7 +854,7 @@ pub fn new_server_cluster(id: u64, count: usize) -> Cluster<ServerCluster, Rocks
 pub fn new_incompatible_server_cluster(
     id: u64,
     count: usize,
-) -> Cluster<ServerCluster, RocksEngine> {
+) -> Cluster<ServerCluster<RocksEngine>, RocksEngine> {
     let pd_client = Arc::new(TestPdClient::new(id, true));
     let sim = Arc::new(RwLock::new(ServerCluster::new(Arc::clone(&pd_client))));
     Cluster::new(
@@ -866,7 +871,7 @@ pub fn new_server_cluster_with_api_ver(
     id: u64,
     count: usize,
     api_ver: ApiVersion,
-) -> Cluster<ServerCluster, RocksEngine> {
+) -> Cluster<ServerCluster<RocksEngine>, RocksEngine> {
     let pd_client = Arc::new(TestPdClient::new(id, false));
     let sim = Arc::new(RwLock::new(ServerCluster::new(Arc::clone(&pd_client))));
     Cluster::new(
@@ -879,14 +884,21 @@ pub fn new_server_cluster_with_api_ver(
     )
 }
 
-pub fn must_new_cluster_and_kv_client() -> (Cluster<ServerCluster, RocksEngine>, TikvClient, Context)
-{
+pub fn must_new_cluster_and_kv_client() -> (
+    Cluster<ServerCluster<RocksEngine>, RocksEngine>,
+    TikvClient,
+    Context,
+) {
     must_new_cluster_and_kv_client_mul(1)
 }
 
 pub fn must_new_cluster_and_kv_client_mul(
     count: usize,
-) -> (Cluster<ServerCluster, RocksEngine>, TikvClient, Context) {
+) -> (
+    Cluster<ServerCluster<RocksEngine>, RocksEngine>,
+    TikvClient,
+    Context,
+) {
     let (cluster, leader, ctx) = must_new_cluster_mul(count);
 
     let env = Arc::new(Environment::new(1));
@@ -898,14 +910,22 @@ pub fn must_new_cluster_and_kv_client_mul(
 }
 pub fn must_new_cluster_mul(
     count: usize,
-) -> (Cluster<ServerCluster, RocksEngine>, metapb::Peer, Context) {
+) -> (
+    Cluster<ServerCluster<RocksEngine>, RocksEngine>,
+    metapb::Peer,
+    Context,
+) {
     must_new_and_configure_cluster_mul(count, |_| ())
 }
 
 fn must_new_and_configure_cluster_mul(
     count: usize,
-    mut configure: impl FnMut(&mut Cluster<ServerCluster, RocksEngine>),
-) -> (Cluster<ServerCluster, RocksEngine>, metapb::Peer, Context) {
+    mut configure: impl FnMut(&mut Cluster<ServerCluster<RocksEngine>, RocksEngine>),
+) -> (
+    Cluster<ServerCluster<RocksEngine>, RocksEngine>,
+    metapb::Peer,
+    Context,
+) {
     let mut cluster = new_server_cluster(0, count);
     configure(&mut cluster);
     cluster.run();
@@ -921,8 +941,12 @@ fn must_new_and_configure_cluster_mul(
 }
 
 pub fn must_new_and_configure_cluster_and_kv_client(
-    configure: impl FnMut(&mut Cluster<ServerCluster, RocksEngine>),
-) -> (Cluster<ServerCluster, RocksEngine>, TikvClient, Context) {
+    configure: impl FnMut(&mut Cluster<ServerCluster<RocksEngine>, RocksEngine>),
+) -> (
+    Cluster<ServerCluster<RocksEngine>, RocksEngine>,
+    TikvClient,
+    Context,
+) {
     let (cluster, leader, ctx) = must_new_and_configure_cluster(configure);
 
     let env = Arc::new(Environment::new(1));
@@ -934,13 +958,20 @@ pub fn must_new_and_configure_cluster_and_kv_client(
 }
 
 pub fn must_new_and_configure_cluster(
-    configure: impl FnMut(&mut Cluster<ServerCluster, RocksEngine>),
-) -> (Cluster<ServerCluster, RocksEngine>, metapb::Peer, Context) {
+    configure: impl FnMut(&mut Cluster<ServerCluster<RocksEngine>, RocksEngine>),
+) -> (
+    Cluster<ServerCluster<RocksEngine>, RocksEngine>,
+    metapb::Peer,
+    Context,
+) {
     must_new_and_configure_cluster_mul(1, configure)
 }
 
-pub fn must_new_cluster_and_debug_client() -> (Cluster<ServerCluster, RocksEngine>, DebugClient, u64)
-{
+pub fn must_new_cluster_and_debug_client() -> (
+    Cluster<ServerCluster<RocksEngine>, RocksEngine>,
+    DebugClient,
+    u64,
+) {
     let (cluster, leader, _) = must_new_cluster_mul(1);
 
     let env = Arc::new(Environment::new(1));
