+++ components/test_raftstore-v2/src/util.rs
@@ -11,7 +11,7 @@ use futures::Future;
 use kvproto::{kvrpcpb::Context, metapb, raft_cmdpb::RaftCmdResponse};
 use raftstore::Result;
 use rand::RngCore;
-use server::server2::ConfiguredRaftEngine;
+use server::common::ConfiguredRaftEngine;
 use tempfile::TempDir;
 use test_raftstore::{new_get_cmd, new_put_cf_cmd, new_request, Config};
 use tikv::{
@@ -163,7 +163,7 @@ pub fn configure_for_lease_read_v2<T: Simulator<EK>, EK: KvEngine>(
 }
 
 pub fn wait_for_synced(
-    cluster: &mut Cluster<ServerCluster, RocksEngine>,
+    cluster: &mut Cluster<ServerCluster<RocksEngine>, RocksEngine>,
     node_id: u64,
     region_id: u64,
 ) {
